Block ponential security holes in code loader
Make sure that CNTXT_ITEM opcode's jump target check comes *before* the end of
the bytecode program instead of the previous one past the end.
vm::MAX_OPCODE is used to check if an incomming opcode value is in the range of
legitemate *on disk* opcodes.  It was being defined to include the internal
TEMP_COPY instruction in that set, which would have potentially allowed
internal only TEMP_COPY instructions to be read from the on disk instruction
stream.