Promote unchanging args on Machine::run to the Machine class
All Slots in a SlotMap belong to the same segment, therefore it makes sense for
the SlotMap to know it's segment since both are often needed together.
As both the SlotMap and segment objects never change for the lifetime of the
Machine give Machine a reference to the SlotMap.  These changes remove the need
to repeatedly pass arguments for those objects on every call to Machine::run