Remove m_before and m_after from CharInfo. Add non features to hacking.txt
m_before and m_after in CharInfo were not being used for anything useful and now their values are calculated
as a separate processing function on the segment and the information is not held in the CharInfo. Therefore
remove these member variables and the various functions and parameters so associated. CharInfo is now rather
lightweight!