Remove redundant asserts from reported error conditions in TtfUtil
In GetTableDirInfo and CheckTable there are asserts that reproduce runtime
error testing and cause client to abort() in a recoverable error situation.
These asserts are ignored in Release builds but present in Debug, and
RelWIthDebInfo builds casuing them to not always show up depending on how
the version under test was built.