Improve the sanity checking in galt iterators and GlyphFace ctor
When the glat input iterator will overrun the rhs of a == comparision assume we
have an invalid iterator and always report equal. This will halt any loops
before wandering past the end.
Added sanity checks to confirm that the glyphs glat range is smaller than the
worst case packing for m_numAttrs.
Check that a sparse array doesn't have more values than m_numAttrs, destroy it
and initialise an emtpy array if it does.