Replaced all trace WORDS_BIGENDIAN and with a endian agnostic decoder
Replaced the swapXX and readXX functions with the Endian.h template versions
that are CPU byte order agnostic.
Modified gr2fonttest to not need to know the CPU byte order
Add the endian test harness so any byte order related errors will fail that
test in addition to the typical mysterious failures in other tests.