Quit executing graphite if the VM finishes with an error status
Trying to run code on a VM that is has an error status is now an
assertable condition.  Once a Mahcine encounters an error condition
it is effectivly wedged.
Refactor status information into the Machine object and out of
reference parameters.
Move vm::Code into vm::Machine as an inner class to grant it access
to a Machine's status and made vm::Machine::run private.
Fix up all call sites and added early quit code if a Machine error
status is detected.