Incorprate imporvements and fixes from Martin to the schema
Add an optional "bidi" level member to Slot.
Change the Point schema type from an object to a 2 element array for
easier use in the debugger.
Remove a specific schema for Advance and reuse the Point schema.
Grouped parent attachment point and id together and made the
attachment point an offset from the parent's origin in anticipation
of similar simplification in the graphite engine.
Renamed Slot member "attachment" to "children" to better reflect the
tree relationship between attached slots.
Added an "input" member to RuleEvent as an offset and length on it's
Pass' slots.
Rename RuleEvent's "slots" member to "output" for clarity.
Changed the maximum for CharInfo's "unicode" member from hex to
decimal as apparently that's not proper JSON.
Change CharInfo's "breakWeight" from an enumerated type to an integer
between -50..50 as a CharInfo can take on values other than those
enumerated in graphite2/Segment.h:gr_break_weight