Fix misaligned load on Sparc CPU causing segfault
Transform all uses of {be,le}::swap of the forms:
  swap(*(T *)(pointer expr)) -> peek<T>(pointer expr)
  swap(<array of T>[i])      -> peek<T>(<array of T> + i)
The calculated address derefernced before calling swap may be missaligned and in these cases it is possible for compiler to not recognise this (especially where the address has been cast from void * or byte *) and emit a misaligned load. As peek (and by extension read) access memory as bytes the are immune from alignment issues.
All other uses of swap are (e.g. on packed structure members) is fine as the compiler can recognise a misaligned load and emit the correct instructions to compensate.