Be nice to the compiler
Wrap src/inc/debug.h in GRAPHITE2_NTRACING checks. Should not be required as none of the functions are being called when GRAPHITE2_NTRACING is in effect anyway, but this means less work for the compiler.  Plus any mistakes will show up in the compilation phase rather than the link phase.