Avoid possible undefined behaviour if `Font::Table` is ever subclassed
`this->~Table()` is a possible source of undefined behaviour in any
future derived classes of `Font::Table`. This should have at least been
`this->~Table::Table()`, but we replace it with an explicit call to
`release()` since that's all it does anyway.
Also make `Font::Table` move only and delete copying ctor and assignment
to enforce unique_ptr like behaviour. We would use unique_ptr itself but
it requires runtime support on gcc & clang. Closes Mozilla #1448175